+++ src/features/whatsnew/whatsNew.js
@@ -38,104 +38,22 @@ ___
 
 The complete changelog can be found [here](https://github.com/ConsenSys/vscode-solidity-auditor/blob/master/CHANGELOG.md). 
 
-## v0.2.0 - 💖 Happy Valentine's Days, spread some love! 💖
+## v0.2.0 - 🍂🍁🍃 Atumn Is Here! 🎃👻
 
 - fix: includes parser performance optimizations with [solidity-workspace@v0.2.0](https://github.com/tintinweb/solidity-workspace/releases/tag/v0.2.0) (🙏 @vquelque)
 - update: completely refactored function signature computation - #127
-- fix: performance optimizations (debounce) - #127
+- fix: performance optimizations - #127
 - new: command \`solidity-va.tools.function.signatures.forWorkspace\` to show all function signatures for files in the workspace - #127
 - fix: \`ftrace\` error in cockpit view due to non-existent filepath being passed to surya - #127
 - fix: draw.io/plantuml mistakenly named \`receive()|fallback()\` as \`constructor\` - #127
 - fix: function signature extraction only returns \`public|external\` interfaces - #127
 - fix: external function call list in cockpit view faile to resolve one-liner function declarations - #127
-
-## v0.1.5 - 🏂
-
-- fix: find all references not working - #119
-
-## v0.1.4 - 🍁
-
-- fix: incompatibility with vscode update v1.72.0 - Extension "[...] has provided an invalid tree item." - #114
-- new: optionally disable highlighting in the overview ruler - #115
-
-<img width="440" alt="image" src="https://user-images.githubusercontent.com/2865694/194864316-88b89bf4-331d-43b4-bb0a-324cdcee99da.png">
-
-<img width="185" alt="image" src="https://user-images.githubusercontent.com/2865694/194864195-a1449bb3-9c04-43bb-a011-8d2faf8ffb5a.png"> <img width="181" alt="image" src="https://user-images.githubusercontent.com/2865694/194864429-5d8c42d4-e45b-406d-ad66-e207718aced0.png">
-
-- fix: more consistent highlighting when clicking on an item in the cockpit ExternalCalls/PublicStatechanging views - #110
-    - clicking on a function name highlights the first line of the function instead of the full function block in the editor
-    - clicking on an external call/modifier highlights the invocation in the editor
-
-- fix: semantic highlighting for supercall and modifier arguments - #112 #117
-
-## v0.1.3 - 🧸
-
-- new: customize semantic highlighting mode #105 #108
-    - \`color and symbol\` - default (decorate + "arrow")
-    - \`color only\` - only decorate identifiers declared in the function argument scope 
-    - \`symbol only\` - only show an "arrow" next to an identifier declared in the function argument scope
-
-<img width="725" alt="image" src="https://user-images.githubusercontent.com/2865694/180291604-f1a3be47-5aaa-41d4-b734-13dfd813d8ff.png">
-
-- fix: uml - fix null-deref when parsing 'using-for *' #106 #107
-- update: configurable trailing "arrow" for semantic function argument highlighting #104 #100
-
-
-## v0.1.2
-
-- new: decorate/highlight immutable state vars (<span style="color:#9932cc">![#9932cc](https://via.placeholder.com/15/9932cc/000000?text=+)</span>) - #97 (thanks @RomiRand)
-<img width="381" alt="image" src="https://user-images.githubusercontent.com/2865694/163415669-7d45d698-2be1-49a3-80b4-c6e2ed861d11.png">
-
-- update: dependencies (surya@0.4.6)
-- new: configuration option to enable the \`surya.mdreport\` "negative Modifiers" feature, see [surya#162](https://github.com/ConsenSys/surya/pull/162)
-    - enabling \`solidity-va.tools.surya.option.negModifiers\` will list all modifiers observed in the file with the ones that are not being used with the listed method being  ~~striked-through~~
-
-![image](https://user-images.githubusercontent.com/2998191/155733325-7a6187b8-e63e-4410-a312-aa2a1c940e31.png)
-
-  Note that the report can be generated either via the \`report\` codelense or by selecting files in the \`Solidity Visual Developer View → right-click → Surya: generate report\`.
-
-<img width="401" alt="image" src="https://user-images.githubusercontent.com/2865694/163411802-49e91a8d-df9e-44ca-8c62-23510d7c9a4a.png">
-
-<img width="398" alt="image" src="https://user-images.githubusercontent.com/2865694/163412288-20e621df-b715-4074-b8f8-033a4b758002.png">
-
-
-- fix: typos & links to placeholder[.]com - #93 #91 (thanks @almndbtr)
-
-## v0.1.1 - ❄️🎄🏂🎄❄️
-
-- fix: type resolving and declaration link for inherited statevars
-- update: move language specific logic to a web compatible extension
-    - https://github.com/tintinweb/vscode-solidity-language (https://marketplace.visualstudio.com/items?itemName=tintinweb.vscode-solidity-language)
-    - The language ships with three security centered Color Themes that can be selected from the **Solidity Language & Themes (only)** extension page or \`Code → Preferences → Color Themes\` 
-    
-    ![image](https://user-images.githubusercontent.com/2865694/145625639-a54178d1-5cee-45b6-bf10-44da09875f0e.png)
-
-- update: support for latest solidity parser (extension + surya) #84
-
-## v0.1.0 - with a lot of new features 🥳
-
-- new: 🥳 major parser refactoring #67 and we published the parser as standalone lib "[solidity-workspace](https://github.com/tintinweb/solidity-workspace)"
-- new: ⚠️ annotations for potential external calls (yellow gutter icon)
-    - best effort - might miss some external calls depending on whether it is possible to easily resolve them (e.g. accessing addresses in nested structures/mappings).
-- new: cockpit view that lists external calls in the currently selected contract (click into a contract in the editor for the view to update)
-
-  ![image](https://user-images.githubusercontent.com/2865694/122222447-90933880-ceb2-11eb-91c3-c59549d40c8c.png)
-
-- new: we can now resolve inherited names (hover: declaration link)
-  
-  ![image](https://user-images.githubusercontent.com/2865694/120014274-26d5ec00-bfe2-11eb-99f7-64d4a57277a0.png)
-
-- new: we now decorate identifiers that are storage references (treating them like state-vars)
-- new: unit-test stub/template for Hardhat/Ethers #70 (\`preferences → Settings → Solidity Visual Developer: solidity-va.test.defaultUnittestTemplate\`)
-- new: (debug) option to enable/disable stacktraces for parser errors (\`preferences → Settings → Solidity Visual Developer: solidity-va.debug\`)
-- new: show codelenses (inline actions) for abstract contracts
-- new: customize which codelenses to show or hide (\`preferences → Settings → Solidity Visual Developer: solidity-va.codelens.*\`) #76
-- new: expose new command \`solidity-va.surya.graphThis\` #76
-- new: use internal ("dumb" lexical) flattener by default. Optionally allow to select \`truffle-flattener\` (\`preferences → Settings → Solidity Visual Developer: solidity-va.flatten.mode\`)
-- update: enable \`draw.io csv export\` codelens by default
-- fix: misplaced decoration when document changes
-- fix: function selector is incorrect if there's a comment in the function signature definition #68
-- update: code cleanup; refactored decoration logic and moved it to its own submodule
+- fix: inheritance hover showing "object" instead of inherited contract name
+- update: dependencies (solidity parser / solidity workspace)
+- fix: constructor/fallback showing up as 'null' in cockpit
+- fix: redecorate editor only if changes were detected (performance)
+- fix: abort decoration if content hash not found in cache (keep current decoration if file is unparseable)
+- fix: cockpit view for overridden functions (#138)
 
 
 <sub>